D-Bus testing: allow SYNCEVOLUTION_DEBUG in some Cmdline tests, but not all
Always unsetting SYNCEVOLUTION_DEBUG removed an important debugging
method: SYNCEVOLUTION_DEBUG is useful in cases where output
redirection and/or handling is broken.

Therefore this commit removes the forced unsetting of
SYNCEVOLUTION_DEBUG again and instead let's at least
TestCmdline.testConfigureTemplates pass with and without the
additional debug output.

Other tests fail when it is set, so the right way to invoke
test-dbus.py is without it.