D-Bus testing: added logging facility for Python test driver
Normal print statements are not very useful because they are
out-of-band (not tied to either the D-Bus log, the SyncEvolution
output or the failure message of a failed test).

The new logging.log() (for simple string) and logging.printf() (for
string formatting with arbitrary arguments) get the output included in
the .dbus.log of a test by sending out a signal on a dummy
object. Search for =log in the .dbus.log:

signal sender=:1.1194 -> dest=(null destination) serial=35 path=/test/dbus/py; interface=t.d.p; member=log
   string "killing syncevo-dbus-server with pid 7851"