D-Bus testing: don't configure password for local sync
Specifying - for password in a local sync doesn't make sense: the
password is not needed. Apparently the
TestSessionAPIsDummy.testAutoSyncLocalSuccess test worked because the
password was never checked. There's no guarantee that this is always
the case and in fact, the fork/exec version of syncevo-dbus-server did
ask for the password.

Better fix the test so that it does not depend on passwords being
ignored for local sync.