ForkExec: added parent connection monitoring
A parent already could detect when the child died, but the child
wasn't notified when the parent died or (more important) closed the
connection.

Now it is possible to detect loss of connection, which includes
"parent died" but also the weaker "parent closed its side of the
communication". The idea is the same as in the handling of pending
operations in syncevo-dbus-helper: invoke a method and keep the method
call open forever. When the connection is lost, D-Bus will invoke the
error callback of the pending method.