D-Bus testing: added test for completing sync without syncevo-dbus-server
One of the goals of the fork/exec rewrite is to let a sync complete
even when syncevo-dbus-server dies. The new TestLocalSync.testNoParent
covers that; to pass it, the server had to be changed so that its
signal delivery to the D-Bus server became optional (= no exception
thrown when it fails).