syncevo-dbus-helper: log SuspendFlags at DEBUG level
The "aborting" messages created by SuspendFlags are of no interest
to the user and thus shouldn't be logged at INFO level. Worse, if they
get logged while output redirection is off, then they go to stdout and from
there to syncevo-dbus-server, which treats any output as ERROR because
it cannot know better.