GLibSupport: improved SE_GOBJECT_TYPE
Include boost/shared_ptr.hpp, it is needed for the macro, which is
used inside the header.

Added implicit cast to plain pointer, without making it const. This is
fairly safe, because the usual semantic of C functions taking such
pointers is to assume that the caller retains ownership (in contrast
to GError, which often passes ownership!).