Akonadi: don't auto-start daemon
A normal KDE user should already have the akonadi server running.
Others probably don't want it to be started, for example as part
of --print-databases.

Therefore auto-starting Akonadi is now disabled. Attempts to use
the Akonadi sources results in an error message with a hint
about akonadictl when Akonadi is not already running.

The same problem is less relevant for EDS (but still exists to
some extend): here listing databases "merely" starts gconfd.