D-Bus session: updated Akonadi startup
Pre-starting Akonadi for test-dbus.py doesn't work because of differences
in the environment of Akonadi server and client (XDG root). Disabled
it for D-Bus testing, which must work without Akonadi (and cannot
test it...).

Also changed the sleep statement so that dbus-session.sh sleeps after
starting either EDS or Akonadi.