logging: control libsynthesis console output via SYNCEVOLUTION_DEBUG
libsynthesis console is logged with a "SYSYNC: " prefix for each line
if (and only if) SYNCEVOLUTION_DEBUG is set to a value of 4 or higher.
Showing it by default is not useful (because it can be quite verbose,
in particular when a server triggers decoding error messages in SyncML
TK, like Funambol) and using the normal loglevel doesn't help because
that loglevel is not available outside of sync sessions.