sync: explicitly ask for refresh-from-server (helps with Funambol)
libsynthesis has traditionally implemented "refresh-from-server" as
"delete local data" plus "slow" sync. This is more compatible, because
some servers did not support "refresh-from-server".

But it has the downside that the server cannot know that the client
won't send any data, and Funambol now only allows one slow sync before
blocking the next one for a certain period of time. Probably this is
done to prevent excessive resource usage by badly behaving clients.

SyncEvolution now uses a new configure option in libsynthesis which
enables the use of "refresh-from-server". This is set unconditionally;
all modern SyncML servers are expected to support it.