local sync: fixed hanging syncevo-local-sync
Somehow syncevo-local-sync got stuck waiting for its parents reply to
a D-Bus method call when the parent had already quit. The expected
error response did not occur in this case.

Fixed by watching whether the parent is around while waiting for the
response. Also got rid of the event loop start/stop in favor of
g_main_context_iteration(). That is more explicit about the reason for
stopping the waiting.