WebDAV: improved VJOURNAL -> plain/text conversion
When a memo in VJOURNAL format from the CalDAV backend gets converted
to plain text *and* the description starts with summary plus newline,
then only the description is used in the plain text version, to avoid
duplicating the summary.

Such entries are produced by SyncEvolution when importing plain text
into EDS. The conversion in the CalDAV backend uses a slightly
different logic and strips the first line from the incoming text
instead of just copying it.

The main reason for this change is that the new mechanism (based
libsynthesis text profile) makes it easier to implement the conversion
that way.