command line: improved support for aborting while checking sources
A --configure operation checks whether backends are usable before
enabling them. This can take a while or fail for every single source
after a long timeout, in particular when the backend needs to contact
a remote server and that server was not configured correctly or is
dead (WebDAV!).

Now the command line checks for abort requests in the loop over
sources.