testing: use "refresh-from-server"
client-test avoided "refresh-from-server" in favor of delete+slow sync
because the Funambol SyncML did a literal "refresh-from-server" sync.
libsynthesis is smarter and does the delete+slow sync or literal
refresh sync, depending on the configuration for the peer.

Changed because Funambol's OneMedia throttles clients with 417 which
ask for slow syncs.